feat: package exports
