+++ package.json
@@ -2,14 +2,37 @@
   "name": "@govtechsg/tradetrust-utils",
   "version": "0.1.0",
   "description": "",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/src/index.d.ts",
+  "types": "./dist/types/index.d.ts",
+  "typesVersions": {
+    "*": {
+      ".": [
+        "./dist/types/index.d.ts"
+      ],
+      "constants/*": [
+        "./dist/types/constants/*.d.ts"
+      ]
+    }
+  },
+  "exports": {
+    ".": {
+      "require": "./dist/cjs/index.js",
+      "import": "./dist/esm/index.js"
+    },
+    "./constants/*": {
+      "require": "./dist/cjs/constants/*.js",
+      "import": "./dist/esm/constants/*.js"
+    }
+  },
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "build": "npm run clean && npm run build:cjs && npm run build:esm && npm run build:type",
+    "build": "npm run clean && npm run build:cjs && npm run build:esm && npm run build:types",
     "build:cjs": "tsc --module commonjs --outDir dist/cjs --project ./tsconfig.prod.json",
     "build:esm": "tsc --module es2015 --outDir dist/esm --project ./tsconfig.prod.json",
-    "build:type": "tsc -d --emitDeclarationOnly --outDir dist/types/src",
+    "build:types": "tsc --declaration --emitDeclarationOnly --outDir dist/types --project ./tsconfig.prod.json",
     "clean": "rm -rf dist/",
     "test:ci": "jest --runInBand",
     "test": "NODE_OPTIONS=--max-old-space-size=2048 jest",
@@ -43,7 +66,7 @@
     "prettier": "^2.5.1",
     "semantic-release": "^18.0.1",
     "ts-jest": "^27.1.0",
-    "typescript": "^4.5.2"
+    "typescript": "^4.9.4"
   },
   "publishConfig": {
     "access": "public"
