+++ src/constants/network.ts
@@ -1,0 +1,5 @@
+export type networkName = "local" | "mainnet" | "matic" | "maticmum" | "goerli" | "sepolia";
+
+export type networkType = "production" | "test" | "development";
+
+export type networkCurrency = "ETH" | "MATIC";
