+++ src/index.ts
@@ -1,5 +1,5 @@
 import * as CONSTANTS from "./constants/VerificationErrorMessages";
-export * from "./constants/SupportedNetworks";
+export * from "./constants/supportedChains";
 export * from "./analytics";
 export * from "./fragments";
 export { CONSTANTS };
